Lock down queries, mutations and subscription

+++ src/resolvers/Mutation.js
@@ -41,14 +41,17 @@ const Mutation = {
     }
   },
 
-  async deleteUser(parent, args, { prisma }, info) {
-    return prisma.mutation.deleteUser({where: {id: args.id}})
+  async deleteUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+    return prisma.mutation.deleteUser({where: {id: userId}})
   },
 
-  updateUser(parent, args, { prisma }, info) {
+  updateUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+    
     return prisma.mutation.updateUser({
       where: {
-        id: args.id
+        id: userId
       }, 
       data: args.data
     }, info)
@@ -67,15 +70,54 @@ const Mutation = {
         }}}, info)
   },
 
-  deletePost(parent, args, { prisma, pubsub }, info) {
-    return prisma.mutation.deletePost({
+  // check user exists because there is no where -> author. Look at :4466, 
+  async  deletePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
+    const postExists = prisma.exists.Post({
+      id: args.id, 
+      author: {
+        id: userId
+      }
+    })
+
+    if (!postExists) throw new Error('Unable to delete post')
+
+    return await prisma.mutation.deletePost({
       where: {
         id: args.id
       }   
     }, info)
   },
 
-  updatePost(parent, args, { prisma, pubsub }, info) {
+  async updatePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
+    const postExist = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if (!postExist) throw new Error('Unable to update')
+
+    const post = await prisma.query.post({
+      where: {
+       id: args.id
+      }
+    })
+
+    if (post.published && args.data.published === false) {
+      await prisma.mutation.deleteManyComments({
+        where: {
+          post_id: {
+            id: args.id
+          }
+        }
+      })
+    }
+
     return prisma.mutation.updatePost({
       where: {
         id: args.id,
@@ -84,13 +126,22 @@ const Mutation = {
     }, info)
   },
 
-  createComment(parent, args, { prisma, pubsub }, info) {
+  async createComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
+    const postPublished = await prisma.exists.Post({
+      id: args.data.post_id,
+      published: true
+    })
+
+    if(!postPublished) throw new Error('Post is not published')
+
     return prisma.mutation.createComment({
       data: {
         text: args.data.text,
         author_id: {
           connect: {
-            id: args.data.author_id
+            id: userId
           }
         },
         post_id: {
@@ -102,7 +153,19 @@ const Mutation = {
     })
   },
 
-  deleteComment(parent, args, { prisma, pubsub }, info) {
+  async deleteComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+    console.log(userId)
+
+    const commentExist = await prisma.exists.Comment({
+      id: args.id,
+      author_id: {
+        id: userId
+      }
+    })
+
+    if (!commentExist) throw new Error('Unable to delete')
+
     return prisma.mutation.deleteComment({
       where: {
         id: args.id
@@ -110,7 +173,18 @@ const Mutation = {
     }, info)
   },
 
-  updateComment(parent, args, { prisma, pubsub }, info) {
+  async updateComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
+    const commentExist = await prisma.exists.Comment({
+      id: args.id,
+      author_id: {
+        id: userId
+      }
+    })
+
+    if (!commentExist) throw new Error('Unable to delete')
+
     return prisma.mutation.updateComment({
       data: args.data,
       where: {
